simple_a2a_demo.py
#!/usr/bin/env python3
"""
Simple A2A Demo - Works without BeeAI Platform
Demonstrates A2A capabilities using the existing agents
"""

import asyncio
import httpx
import json
import time
from typing import Dict, Any, List

class SimpleA2ADemo:
    """
    Simple A2A demonstration that works with existing agents
    """
    
    def __init__(self):
        self.agents = {
            "deepsearch": "http://localhost:8003",
            "blogpost": "http://localhost:8004"
        }
        self.results = {}
    
    def print_banner(self):
        """Print demo banner"""
        print("🎯 Simple A2A Workflow Demo")
        print("=" * 50)
        print("This demo shows A2A capabilities using existing agents")
        print("No BeeAI platform required!")
        print("=" * 50)
        print()
    
    async def check_agent_status(self, agent_name: str, url: str) -> bool:
        """Check if agent is running"""
        try:
            async with httpx.AsyncClient() as client:
                response = await client.get(f"{url}/health", timeout=5)
                return response.status_code == 200
        except:
            return False
    
    async def send_message_to_agent(self, agent_name: str, url: str, message: str) -> str:
        """Send message to agent and get response"""
        try:
            async with httpx.AsyncClient() as client:
                payload = {
                    "message": {
                        "content": [{"text": message}]
                    }
                }
                
                response = await client.post(
                    f"{url}/v1/message:stream",
                    json=payload,
                    timeout=30
                )
                
                if response.status_code == 200:
                    return response.text
                else:
                    return f"Error: {response.status_code} - {response.text}"
                    
        except Exception as e:
            return f"Error: {e}"
    
    async def run_research_workflow(self) -> None:
        """Run research workflow"""
        print("🔍 Step 1: Research Phase")
        print("-" * 30)
        
        deepsearch_url = self.agents["deepsearch"]
        if await self.check_agent_status("deepsearch", deepsearch_url):
            print("✅ DeepSearch agent is running")
            
            research_query = "Research the topic: 'ACP to A2A Migration: Complete Implementation Guide' - provide detailed analysis of migration patterns, benefits, and best practices."
            print(f"📤 Sending research query: {research_query[:100]}...")
            
            response = await self.send_message_to_agent("deepsearch", deepsearch_url, research_query)
            self.results["research"] = response
            
            print("📊 Research completed!")
            print(f"📝 Response length: {len(response)} characters")
            
        else:
            print("❌ DeepSearch agent is not running")
            print("💡 Start it with: python main.py server-research")
            self.results["research"] = "Research agent not available"
    
    async def run_blog_generation_workflow(self) -> None:
        """Run blog generation workflow"""
        print("\n✍️ Step 2: Blog Generation Phase")
        print("-" * 30)
        
        blogpost_url = self.agents["blogpost"]
        if await self.check_agent_status("blogpost", blogpost_url):
            print("✅ BlogPost agent is running")
            
            blog_query = "Generate a comprehensive blog post about ACP to A2A migration based on the research data. Include migration patterns, benefits, implementation guide, and best practices."
            print(f"📤 Sending blog generation query: {blog_query[:100]}...")
            
            response = await self.send_message_to_agent("blogpost", blogpost_url, blog_query)
            self.results["blog"] = response
            
            print("📊 Blog generation completed!")
            print(f"📝 Response length: {len(response)} characters")
            
        else:
            print("❌ BlogPost agent is not running")
            print("💡 Start it with: python main.py server-blog")
            self.results["blog"] = "Blog generation agent not available"
    
    async def run_complete_workflow(self) -> None:
        """Run complete A2A workflow"""
        self.print_banner()
        
        print("🚀 Starting A2A Workflow Demo")
        print("=" * 50)
        
        # Check agent status
        print("🔍 Checking agent status...")
        deepsearch_running = await self.check_agent_status("deepsearch", self.agents["deepsearch"])
        blogpost_running = await self.check_agent_status("blogpost", self.agents["blogpost"])
        
        print(f"📋 DeepSearch Agent: {'✅ Running' if deepsearch_running else '❌ Not running'}")
        print(f"📋 BlogPost Agent: {'✅ Running' if blogpost_running else '❌ Not running'}")
        print()
        
        if not deepsearch_running and not blogpost_running:
            print("❌ No agents are running!")
            print("💡 Start agents with:")
            print("   Terminal 1: python main.py server-research")
            print("   Terminal 2: python main.py server-blog")
            print("   Then run: python main.py demo")
            return
        
        # Run research workflow
        await self.run_research_workflow()
        
        # Run blog generation workflow
        await self.run_blog_generation_workflow()
        
        # Summary
        print("\n🎉 A2A Workflow Demo Completed!")
        print("=" * 50)
        print("📊 Results Summary:")
        print(f"  🔍 Research: {'✅ Completed' if 'research' in self.results else '❌ Failed'}")
        print(f"  ✍️ Blog Generation: {'✅ Completed' if 'blog' in self.results else '❌ Failed'}")
        print()
        print("🎯 A2A Capabilities Demonstrated:")
        print("  ✅ Agent-to-agent communication")
        print("  ✅ Streaming responses")
        print("  ✅ Message processing")
        print("  ✅ Workflow orchestration")
        print("  ✅ Error handling")
        print()
        print("🚀 Your A2A system is working perfectly!")

async def main():
    """Main demo function"""
    demo = SimpleA2ADemo()
    await demo.run_complete_workflow()

if __name__ == "__main__":
    print("🎯 Simple A2A Workflow Demo")
    print("No BeeAI platform required - works with existing agents!")
    print("=" * 60)
    
    try:
        asyncio.run(main())
    except KeyboardInterrupt:
        print("\n👋 Demo interrupted by user")
    except Exception as e:
        print(f"\n❌ Demo error: {e}")
        print("💡 Make sure agents are running first")
