pyproject.toml
[project]
name = "acp-migration-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.8.0",
    "azure-identity==1.23.0",
    "azure-storage-blob>=12.27.0",
    "azure-storage-file-datalake>=12.22.0",
    "beautifulsoup4>=4.12.2",
    "beeai-sdk>=0.3.0",
    "crewai-tools[mcp]>=0.75.0",
    "crewai[tools]>=0.201.1",
    "fastapi>=0.104.0",
    "html2text>=2020.1.16",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "mcp[cli]>=1.12.4",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "tenacity>=8.2.3",
    "uvicorn>=0.35.0",
]
